Fixed a memory leak in jar.py
Apparently when initializing multiple instances of the session class, the optional kwargs argument was not cleared. kwargs is initially an empty dict, but is updated to contain a list of None with length corresponding to the number of stars in a session.

If initializing two instances of session, the second would inherit kwargs from the first, and raise and error if the number of stars is different in each one.

Fixed by doing lkwargs = kwargs.copy() at the top of __init__, and working with that instead of kwargs.